# logical: use implicit txns for each INSERT and DELETE