**logical: use implicit txns for each INSERT and DELETE**

This commit introduces another strategy for processing rows. Previously, we would handle a batch of rows within an explicit txn, and this commit makes it so that each row is handled in a separate implicit txn. This strategy is now used by default and can be disabled via newly-added `logical_replication.consumer.use_implicit_txns.enabled` cluster setting. Usage of the implicit txns allows us to hit 1PC down in the KV layer in some cases.

**logical: remove some overhead with explicit txns**

Previously, when using the explicit txns to perform INSERT and DELETE ingestion queries we would use "vanilla" `Txn` method. This method creates a new internal session data on each invocation of the internal executor methods. This commit avoids doing that by supplying the Txn method with the copy of the session data from the session that started the replication stream - that copy will be adjusted to apply some internal executor overrides but otherwise will be unchanged. In other words, even though this commit is not exactly a no-op, it is effectively one.

Epic: CRDB-39063.

Release note: None