@yuzefovich 2024-06-22
First commit is #126051.

The benchmarks show some slowdown because we now create more state on each `ExecParsed` call.
```
name                                 old time/op    new time/op    delta
LastWriteWinsInsert/noPrevValue-8       802µs ± 1%     830µs ± 1%  +3.55%  (p=0.000 n=9+8)
LastWriteWinsInsert/withPrevValue-8    1.18ms ± 1%    1.21ms ± 1%  +2.64%  (p=0.000 n=10+10)

name                                 old alloc/op   new alloc/op   delta
LastWriteWinsInsert/noPrevValue-8       177kB ± 4%     180kB ± 4%  +1.32%  (p=0.019 n=9+9)
LastWriteWinsInsert/withPrevValue-8     242kB ± 1%     244kB ± 0%  +0.80%  (p=0.000 n=9+8)

name                                 old allocs/op  new allocs/op  delta
LastWriteWinsInsert/noPrevValue-8       1.08k ± 1%     1.17k ± 1%  +8.72%  (p=0.000 n=8+8)
LastWriteWinsInsert/withPrevValue-8     1.63k ± 0%     1.73k ± 0%  +5.69%  (p=0.000 n=9+9)
```
We can avoid this by splitting `ProcessRow` into two methods.